shiyan-main/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
volcenginesdkarkruntime
httpx